refactor(main-api): view 패키지 이동
- ActiveDevice , ActiveDeviceViewer -> view package
