+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/device/application/active/view/ActiveDeviceViewer.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.query.device.application.active;
+package com.whoz_in.main_api.query.device.application.active.view;
 
 import com.whoz_in.main_api.query.shared.application.Viewer;
 import java.util.List;
