+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/domain/device/active/ActiveDeviceFilter.java
@@ -5,8 +5,8 @@ import com.whoz_in.domain.device.model.Device;
 import com.whoz_in.domain.device.model.DeviceId;
 import com.whoz_in.domain.network_log.MonitorLog;
 import com.whoz_in.domain.network_log.MonitorLogRepository;
-import com.whoz_in.main_api.query.device.application.active.ActiveDevice;
-import com.whoz_in.main_api.query.device.application.active.ActiveDeviceViewer;
+import com.whoz_in.main_api.query.device.application.active.view.ActiveDevice;
+import com.whoz_in.main_api.query.device.application.active.view.ActiveDeviceViewer;
 import com.whoz_in.main_api.query.device.application.active.event.ActiveDeviceFinded;
 import com.whoz_in.main_api.shared.event.Events;
 import java.util.List;
